#195-binding method 위치 변경
